+++ libs/powderTrain/ui_powderTrain.h
@@ -10,7 +10,7 @@ static unsigned int sbID, sandID, waterID, stoneID, fireID, acidID, woodID;
 // --- Functions ---
 void uiInit_powderTrain();
 int uiGetPressedButton_powderTrain();
-bool uiIsMouseInsideMainWindow_powderTrain(int* mouse_x, int* mouse_y);
+bool uiIsMouseInsideMainWindow_powderTrain(double* mouse_x, double* mouse_y);
 void uiRender_powderTrain();
 void uiAddText_powderTrain();
 void uiCleanUp_powderTrain();
